Refactored to work, also to follow DCoder's review. Added sanity checks for Armory Hospital and Slave Miner. Logics haven't changed yet. Graion's code.
Related to issue #203, related to issue #1523.

git-svn-id: svn://svn.renegadeprojects.com/ares/branches/ft-bounty@1052 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd

+++ src/Ext/TechnoType/Body.h
@@ -119,8 +119,8 @@ public:
 		//#203 Bounty
 		Valueable<bool> Bounty_Message;
 		Valueable<bool> Bounty_FriendlyMessage;
-		Valueable<float> Bounty_Modifier;
-		Valueable<float> Bounty_FriendlyModifier;
+		Valueable<double> Bounty_Modifier;
+		Valueable<double> Bounty_FriendlyModifier;
 		Valueable<bool> ImmuneToBounty;
 		// #1523 also Money Conversion -> Pillage
 		Valueable<bool> Bounty_Pillager;
